requirements.txt
Cython==0.29.1
numpy==1.17.2
pkg-resources==0.0.0
PyOpenGL==3.1.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
scipy==1.3.1
triangle==20190115.3
